feat: implement OurCommunity and Testimonials
