+++ src/pages/ourSupporters/components/Testimonials.tsx
@@ -1,0 +1,23 @@
+import Quotes from "~/components/common/Quotes";
+
+export default function Testimonials() {
+    return <Quotes title="Testimonials" cardData={testimonialData} />;
+}
+
+const testimonialData = [
+    {
+        name: "Mr. David Chua",
+        quote: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque tempor magna aliquet mollis accumsan. Mauris faucibus molestie fringilla. Ut quis sodales lorem, eu dictum turpis. Mauris consectetur lacus non risus sodales, at dictum nisi dapibus. Aenean vehicula eleifend ante luctus feugiat. Phasellus lacinia nisi vitae lectus rhoncus elementum. Morbi vehicula fringilla vulputate.",
+        title: "CEO of the National Youth Council Singapore",
+    },
+    {
+        name: "Mrs. Nancy Tan",
+        quote: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque tempor magna aliquet mollis accumsan. Mauris faucibus molestie fringilla. Ut quis sodales lorem, eu dictum turpis. Mauris consectetur lacus non risus sodales, at dictum nisi dapibus. Aenean vehicula eleifend ante luctus feugiat. Phasellus lacinia nisi vitae lectus rhoncus elementum. Morbi vehicula fringilla vulputate.",
+        title: "Executive Director of XXX",
+    },
+    {
+        name: "Mr. Damien Ng",
+        quote: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque tempor magna aliquet mollis accumsan. Mauris faucibus molestie fringilla. Ut quis sodales lorem, eu dictum turpis. Mauris consectetur lacus non risus sodales, at dictum nisi dapibus. Aenean vehicula eleifend ante luctus feugiat. Phasellus lacinia nisi vitae lectus rhoncus elementum. Morbi vehicula fringilla vulputate.",
+        title: "Director of the YYY",
+    },
+];
