+++ src/pages/ourSupporters/index.tsx
@@ -2,6 +2,7 @@ import PageLayout from "~/components/common/PageLayout";
 import PageTitleSection from "~/components/common/PageTitleSection";
 import Quotes from "~/components/common/Quotes";
 import EventSection from "./components/EventSection";
+import Testimonials from "./components/Testimonials";
 
 export default function ourSupporters() {
     return (
@@ -12,13 +13,13 @@ export default function ourSupporters() {
                     "Amet ex ipsum adipisicing. Consectetur veniam officia pariatur officia in et dolor ex sint consectetur. Dolor aliqua ipsum et labore est. Enim ea tempor quis sint eu non commodo non id anim id non laborum labore. Veniam consequat pariatur duis. Dolor ut laborum in et esse anim commodo eu minim pariatur officia pariatur sit non. Sit aliquip anim sunt eiusmod proident labore Lorem est sit aliqua sunt minim minim occaecat ut."
                 }
             />
-            <EventSection events={supporterData} />
-            <Quotes title="Testimonials" cardData={testimonialData} />
+            <EventSection events={eventData} />
+            <Testimonials />
         </PageLayout>
     );
 }
 
-const supporterData = [
+const eventData = [
     {
         title: "2021 LEAPS PARTNERS",
         description:
@@ -36,23 +37,5 @@ const supporterData = [
             "ABYAteam2.png",
             "ABYAteam3.png",
         ],
-    },
-];
-
-const testimonialData = [
-    {
-        name: "Mr. David Chua",
-        quote: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque tempor magna aliquet mollis accumsan. Mauris faucibus molestie fringilla. Ut quis sodales lorem, eu dictum turpis. Mauris consectetur lacus non risus sodales, at dictum nisi dapibus. Aenean vehicula eleifend ante luctus feugiat. Phasellus lacinia nisi vitae lectus rhoncus elementum. Morbi vehicula fringilla vulputate.",
-        title: "CEO of the National Youth Council Singapore",
-    },
-    {
-        name: "Mrs. Nancy Tan",
-        quote: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque tempor magna aliquet mollis accumsan. Mauris faucibus molestie fringilla. Ut quis sodales lorem, eu dictum turpis. Mauris consectetur lacus non risus sodales, at dictum nisi dapibus. Aenean vehicula eleifend ante luctus feugiat. Phasellus lacinia nisi vitae lectus rhoncus elementum. Morbi vehicula fringilla vulputate.",
-        title: "Executive Director of XXX",
-    },
-    {
-        name: "Mr. Damien Ng",
-        quote: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque tempor magna aliquet mollis accumsan. Mauris faucibus molestie fringilla. Ut quis sodales lorem, eu dictum turpis. Mauris consectetur lacus non risus sodales, at dictum nisi dapibus. Aenean vehicula eleifend ante luctus feugiat. Phasellus lacinia nisi vitae lectus rhoncus elementum. Morbi vehicula fringilla vulputate.",
-        title: "Director of the YYY",
     },
 ];
